+++ teeio-validator/library/pcie_ide_test_lib/CMakeLists.txt
@@ -40,7 +40,8 @@ SET(src_pcie_ide_test_lib
   test_case/test_case_ksetstop_4.c
   test_case/test_case_full.c
   test_case/test_case_full_keyrefresh.c
-  test_case/test_case_spdm_session.c
+  test_case/test_case_spdm_session_1.c
+  test_case/test_case_spdm_session_2.c
   ## test configs
   test_config/test_config_common.c
   test_config/test_config_pcrc.c
